requirements.txt
python-dotenv
langchain 
langchain_community
langchain_groq
langchain_huggingface
faiss-cpu
streamlit